Use template strings and remove XSS (* for the most part) (#1)
* Use template strings and remove XSS (* for the most part)

* The only case that isn't covered are id's which is used
to populate the table elements. The id's will ALWAYS be a
number, since we control them. However, if an attacker were
able to hack the service, they could modify id's. However,
by that point, it is already game over, so I feel that
particular case does not need to be considered in the client
side UI.

* Use JQuery as String Builder instead of using something like JSX

* Fix HTML Structure Errors introduced in the refactor